refactor(workspace): simplify batch upload boundaries
