+++ packages/boring-bash/src/shared/exactBinaryWrite.ts
@@ -1,10 +1,9 @@
-export const EXACT_BINARY_WRITE_POLICIES = ["error", "replace", "skip"] as const;
+export const EXACT_BINARY_WRITE_POLICIES = ["error", "replace"] as const;
 
 export type ExactBinaryWritePolicy = (typeof EXACT_BINARY_WRITE_POLICIES)[number];
 
 export type ExactBinaryWriteOutcome =
   | { status: "written"; path: string }
-  | { status: "skipped"; path: string; reason: "already-exists" }
   | { status: "conflict"; path: string; reason: "already-exists" };
 
 export function isExactBinaryWritePolicy(value: unknown): value is ExactBinaryWritePolicy {
@@ -25,15 +24,8 @@ export function parseExactBinaryWriteOutcome(
   if (outcome.status === "written") {
     return { status: "written", path: outcome.path };
   }
-  if (
-    (outcome.status === "skipped" || outcome.status === "conflict")
-    && outcome.reason === "already-exists"
-  ) {
-    return {
-      status: outcome.status,
-      path: outcome.path,
-      reason: "already-exists",
-    };
+  if (outcome.status === "conflict" && outcome.reason === "already-exists") {
+    return { status: "conflict", path: outcome.path, reason: "already-exists" };
   }
   throw new TypeError("invalid exact binary write outcome status");
 }
