+++ packages/workspace/src/front/bridge/createBridge.ts
@@ -169,7 +169,7 @@ export function createBridge(store: StoreApi): WorkspaceBridge {
     },
 
     async expandToFile(path, opts) {
-      const parsed = expandToFileSchema.safeParse({ path, filesystem: opts?.filesystem, kind: opts?.kind })
+      const parsed = expandToFileSchema.safeParse({ path, filesystem: opts?.filesystem })
       if (!parsed.success) return err("VALIDATION", parsed.error.issues[0].message)
 
       const state = store.getState()
@@ -180,7 +180,6 @@ export function createBridge(store: StoreApi): WorkspaceBridge {
       emit("tree:expand", {
         path: parsed.data.path,
         ...(parsed.data.filesystem ? { filesystem: parsed.data.filesystem } : {}),
-        ...(parsed.data.kind ? { kind: parsed.data.kind } : {}),
       })
       return ok()
     },
