+++ packages/workspace/src/plugins/filesystemPlugin/front/data/__tests__/fetchClient.catalog.test.ts
@@ -6,6 +6,7 @@ const capabilities = {
   list: true,
   search: true,
   write: false,
+  upload: false,
   delete: false,
   move: false,
   mkdir: false,
@@ -17,7 +18,7 @@ describe("FetchClient filesystem catalog", () => {
   it("loads generic roots with auth, credentials, capabilities, and abort signal", async () => {
     const fetchMock = vi.fn(async () => new Response(JSON.stringify({
       filesystems: [
-        { filesystem: "user", label: "Workspace", rootDir: ".", access: "readwrite", capabilities: { ...capabilities, write: true, delete: true, move: true, mkdir: true } },
+        { filesystem: "user", label: "Workspace", rootDir: ".", access: "readwrite", capabilities: { ...capabilities, write: true, upload: true, delete: true, move: true, mkdir: true } },
         { filesystem: "project_alpha", label: "Project alpha", rootDir: "/docs", access: "readonly", capabilities },
       ],
     }), { status: 200 }))
@@ -32,6 +33,16 @@ describe("FetchClient filesystem catalog", () => {
       "https://example.test/api/v1/filesystems",
       expect.objectContaining({ method: "GET", credentials: "include", signal: expect.any(AbortSignal), headers: { Authorization: "Bearer token" } }),
     )
+  })
+
+  it("defaults a missing upload capability to false for mixed-version servers", async () => {
+    const legacyCapabilities = Object.fromEntries(Object.entries(capabilities).filter(([key]) => key !== "upload"))
+    vi.stubGlobal("fetch", vi.fn(async () => new Response(JSON.stringify({
+      filesystems: [{ filesystem: "user", label: "Workspace", rootDir: ".", access: "readwrite", capabilities: legacyCapabilities }],
+    }), { status: 200 })))
+
+    const [root] = await new FetchClient({ apiBaseUrl: "" }).getFilesystems()
+    expect(root?.capabilities.upload).toBe(false)
   })
 
   it("discards malformed and duplicate entries without filesystem-specific inference", async () => {
