+++ packages/workspace/src/plugins/filesystemPlugin/front/data/__tests__/fetchClient.test.ts
@@ -26,7 +26,7 @@ describe("FetchClient", () => {
       status: "conflict",
       path: "src/file.txt",
       reason: "already-exists",
-    }), { status: 409, statusText: "Conflict" }))
+    }), { status: 200 }))
     const client = new FetchClient({ apiBaseUrl: "" })
     const result = await client.writeBinaryFile("src/file.txt", new Blob(["hello"]), { ifExists: "error" })
     expect(result).toEqual({ status: "conflict", path: "src/file.txt", reason: "already-exists" })
@@ -42,7 +42,7 @@ describe("FetchClient", () => {
     const path = " src/file.txt "
     mockFetch
       .mockResolvedValueOnce(new Response(JSON.stringify({ status: "written", path }), { status: 200 }))
-      .mockResolvedValueOnce(new Response(JSON.stringify({ status: "conflict", path, reason: "already-exists" }), { status: 409 }))
+      .mockResolvedValueOnce(new Response(JSON.stringify({ status: "conflict", path, reason: "already-exists" }), { status: 200 }))
     const client = new FetchClient({ apiBaseUrl: "" })
     await expect(client.writeBinaryFile(path, new Blob(["first"]), { ifExists: "error" }))
       .resolves.toEqual({ status: "written", path })
@@ -54,15 +54,15 @@ describe("FetchClient", () => {
   it.each([
     [200, { status: "unknown", path: "src/file.txt" }],
     [200, { status: "written", path: 42 }],
-    [409, { status: "conflict", path: "src/file.txt", reason: "unknown" }],
+    [200, { status: "conflict", path: "src/file.txt", reason: "unknown" }],
   ])("rejects malformed exact-write outcome for HTTP %s", async (statusCode, body) => {
     mockFetch.mockResolvedValue(new Response(JSON.stringify(body), {
       status: statusCode,
-      statusText: statusCode === 409 ? "Conflict" : "OK",
+      statusText: "OK",
     }))
     const client = new FetchClient({ apiBaseUrl: "" })
     await expect(client.writeBinaryFile("src/file.txt", new Blob(["hello"]), { ifExists: "error" }))
-      .rejects.toBeInstanceOf(statusCode === 409 ? FetchError : TypeError)
+      .rejects.toBeInstanceOf(TypeError)
   })
 
   it("rejects a typed conflict for a different exact path", async () => {
@@ -70,10 +70,10 @@ describe("FetchClient", () => {
       status: "conflict",
       path: "src/file.txt",
       reason: "already-exists",
-    }), { status: 409, statusText: "Conflict" }))
+    }), { status: 200 }))
     const client = new FetchClient({ apiBaseUrl: "" })
     await expect(client.writeBinaryFile("src/file.txt ", new Blob(["hello"]), { ifExists: "error" }))
-      .rejects.toBeInstanceOf(FetchError)
+      .rejects.toBeInstanceOf(TypeError)
   })
 
   it("passes abort through binary encoding and transport", async () => {
