+++ packages/workspace/src/shared/plugins/types.ts
@@ -98,7 +98,6 @@ export interface FileTreeRevealRequest {
   path: string
   seq: number
   filesystem?: FilesystemId
-  kind?: "file" | "dir"
 }
 
 export interface LeftTabParams {
